% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_bootstrap.R
\name{render_bootstrap}
\alias{render_bootstrap}
\title{Set knitr bootstrap output functions}
\usage{
render_bootstrap()
}
\description{
This function sets the appropriate knitr output hooks to produce a bootstrap
styled report.
}
\details{
This will allow use of the knitr bootstrap output hooks without the
rmarkdown package.  Note you still need to use pandoc to generate the HTML
document, as the sundown library in the markdown package cannot handle
markdown embedded in HTML tags, which knitr bootstrap uses.
}
