% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knit_bootstrap.R
\name{bootstrap_HTML}
\alias{bootstrap_HTML}
\title{Add the knitrBootstrap html header to an HTML file produced by knitr.}
\usage{
bootstrap_HTML(input, output = NULL)
}
\arguments{
\item{input}{html filename to be wrapped with Bootstrap.}

\item{output}{html filename to output.}
}
\description{
This function includes the knitrBootstrap HTML headers to wrap the knitr
output in bootstrap styled HTML.
}

