% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knit_bootstrap.R
\name{knit_bootstrap}
\alias{knit_bootstrap}
\alias{knit_bootstrap_Rmd}
\title{knit a Rmd file and wrap it in bootstrap styles}
\usage{
knit_bootstrap(input, output = NULL, ...)
}
\arguments{
\item{input}{Rmd input file to knit into HTML}

\item{output}{HTML output file created, if NULL uses the input filename with
the extension changed to .html}

\item{...}{options passed to \code{\link[knitr]{knit}}}
}
\description{
This function includes the knitrBootstrap HTML headers to wrap the knitr
output in bootstrap styled HTML.
}
\seealso{
\code{\link{knit_bootstrap_md}}
}

