% Generated by roxygen2 (4.0.0): do not edit by hand
\name{knit_bootstrap_md}
\alias{knit_bootstrap_md}
\title{knit a md file and wrap it in bootstrap styles}
\usage{
knit_bootstrap_md(input, output = NULL)
}
\arguments{
  \item{input}{md input file to be converted to HTML'}

  \item{output}{HTML output file created, if NULL uses the
  input filename with the extension changed to .html}
}
\description{
This function includes the knitrBootstrap HTML headers to wrap the knitr
output in bootstrap styled HTML.
}
\seealso{
\code{\link{knit_bootstrap}} \code{\link{knit}},
         \code{\link[markdown]{markdownToHTML}}
}

