% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{primerTree}
\alias{primerTree}
\alias{primerTree-package}
\title{\pkg{knitrBootstrap} A framework to create Bootstrap 3 HTML reports from knitr
  Rmarkdown.
Knitr Bootstrap includes a \code{\link{bootstrap_document}} custom
rendering function for use with the rmarkdown
\url{http://http://rmarkdown.rstudio.com/}) package.}
\description{
You can also specify the options in your YAML front-matter, see the
rmarkdown documentation for examples.
}
\examples{
\dontrun{
library(rmarkdown)
library(knitrBootstrap)

render('input.Rmd', bootstrap_document())

#with some options
render('input.Rmd', bootstrap_document(title='Test', highlight='Github',
  theme='Slate', theme.chooser=TRUE, highlight.chooser=TRUE))
}
}

